backend/requirements/base.txt
# Django
Django>=5.0,<5.1
django-cors-headers>=4.3.0
django-filter>=23.5

# Django REST Framework
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.9

# JWT
PyJWT>=2.8.0

# Image Processing
Pillow>=10.1.0

# Storage (AWS S3)
boto3>=1.34.0
django-storages>=1.14.0

# Environment Variables
python-dotenv>=1.0.0

# Date/Time
python-dateutil>=2.8.2

# Utilities
requests>=2.31.0
